Adding welcome screen for ST's

+++ st.php
@@ -3,11 +3,20 @@ include "include.php";
 
 $gameid = $mysqli->real_escape_string($_GET['gameid']);
 $cookie = $mysqli->real_escape_string($_GET['cookie']);
-$mode = $mysqli->real_escape_string($_POST['mode']);
+$gmode = $mysqli->real_escape_string($_GET['gmode']);
+$pmode = $mysqli->real_escape_string($_POST['pmode']);
 $userslist = ($_POST['userslist']);
 
 if ($cookie == "set" && $gameid) {
 	scookie("st","$gameid");
+}
+
+if (!$gmode && !$pmode) {
+	print "Welcome to the ST Arena $username!<br>\n";
+	print "Please choose from the following options:<br><br>\n";
+	if ($_COOKIE['st']) {
+	
+	print "<li><a href=\"st.php?gmode=useredit\">User Editor</a><br>\n";
 }
 
 if (!$_COOKIE['st']) {
@@ -30,20 +39,20 @@ if (!$_COOKIE['st']) {
 	}
 }
 
-if ($_COOKIE['st']) {
+if ($_COOKIE['st'] && $gmode == "useredit") {
 	$cookiegameid = $_COOKIE['st'];
 	$userid = getuseridfromusername($authmysqli,$username);
 	$gameinfo = getgameinfofromgameid($mysqli,$cookiegameid,$userid);
 
 
-	if ($mode == "userset" && $userslist) {
+	if ($pmode == "userset" && $userslist) {
 		// Check to see if their cookie is legit
 		$userid = getuseridfromusername($authmysqli,$username);
 		$cookiegameid = $_COOKIE['st'];
 		$gameinfo = getgameinfofromgameid($mysqli,$cookiegameid,$userid);
 		if (!$gameinfo) {
 			print "Warning! Some cookie douchebaggery is going on here!<br>\n";
-		//	exit;
+			exit;
 		}
 		// First, we'll just delete everyone out of the DB who is a part of that game
 		$deleteaccessusers = $mysqli->query("DELETE FROM access WHERE gameid=\"$cookiegameid\" AND type='U'");
@@ -58,7 +67,7 @@ if ($_COOKIE['st']) {
 	}
 
 
-	print "You are the ST for <u>" . $gameinfo["name"] . "</u><br><br>\n";
+	print "You are editing the user list for <u>" . $gameinfo["name"] . "</u><br><br>\n";
 	print "Please select which people are a part of your game:<br>\n";
 	print "<form action=\"st.php\" method=\"post\">\n";
 	// User list.  First we'll nab from the DB all of the people who are in the game
@@ -79,7 +88,8 @@ if ($_COOKIE['st']) {
 			$out .= "> ".titleCase($users['username'])."<br>";
 			print "$out\n";
 		}
-	print "<input type=\"hidden\" name=\"mode\" value=\"userset\">\n";
+	print "<input type=\"hidden\" name=\"pmode\" value=\"userset\">\n";
+	print "<input type=\"hidden\" name=\"gmode\" value=\"useredit\">\n";
 	print "<input type=\"submit\" value=\"Go\">\n";
 	}
